flask_api/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
docker==6.1.3
dotenv==0.9.9
exceptiongroup==1.3.1
fastapi==0.103.2
Flask==2.2.5
h11==0.14.0
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.6
MarkupSafe==2.1.5
mysql-connector==2.2.9
packaging==24.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==0.21.1
pywin32==308
requests==2.31.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.7
uvicorn==0.22.0
websocket-client==1.6.1
Werkzeug==2.2.3
zipp==3.15.0
